requirements.txt
google-generativeai==0.4.0
python-dotenv==0.19.2
fastapi==0.75.0
uvicorn==0.15.0
pydantic==1.9.0
python-multipart==0.0.5